+++ crochet/_eventloop.py
@@ -213,12 +213,7 @@ class EventualResult(object):
         result.
         """
         if timeout is None:
-            warnings.warn(
-                "Unlimited timeouts are deprecated.",
-                DeprecationWarning,
-                stacklevel=3)
-            # Queue.get(None) won't get interrupted by Ctrl-C...
-            timeout = 2 ** 23
+            raise ValueError("Unlimited timeouts are no longer supported.")
         self._result_set.wait(timeout)
         # In Python 2.6 we can't rely on the return result of wait(), so we
         # have to check manually:
@@ -478,24 +473,6 @@ class EventLoop(object):
         except AttributeError:
             pass
         return result
-
-    def wait_for_reactor(self, function):
-        """
-        DEPRECATED, use wait_for(timeout) instead.
-
-        A decorator that ensures the wrapped function runs in the reactor
-        thread.
-
-        When the wrapped function is called, its result is returned or its
-        exception raised. Deferreds are handled transparently.
-        """
-        warnings.warn(
-            "@wait_for_reactor is deprecated, use @wait_for instead",
-            DeprecationWarning,
-            stacklevel=2)
-        # This will timeout, in theory. In practice the process will be dead
-        # long before that.
-        return self.wait_for(2**31)(function)
 
     def wait_for(self, timeout):
         """